app/requirements.txt
Django==4.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytest
pytest-django
djangorestframework

